pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-rag-qa"
version = "0.1.0"
description = "Medical Document Q&A system using RAG with Qdrant, HuggingFace embeddings, RAGAS evaluation, and FastAPI deployment"
requires-python = ">=3.10"
dependencies = [
    "pymupdf>=1.24",
    "sentence-transformers>=3.0",
    "qdrant-client>=1.9",
    "openai>=1.30",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "ragas>=0.1",
    "langchain>=0.2",
    "langchain-openai>=0.1",
    "langchain-community>=0.2",
    "llama-index>=0.10",
    "gradio>=4.0",
    "mlflow>=2.13",
    "python-dotenv>=1.0",
    "tqdm>=4.66",
    "numpy>=1.26",
    "pandas>=2.2",
    "matplotlib>=3.8",
    "transformers>=4.36.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "reportlab>=4.0",
]

[tool.setuptools.packages.find]
include = ["src*", "app*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-k not slow')",
]
